OdfFormatReader - fix bugs...

+++ ASCOfficeOdfFile/src/docx/xlsx_table_metrics.h
@@ -44,18 +44,18 @@ namespace oox
 class xlsx_table_metrics
 {
 public:
-    xlsx_table_metrics();
-    ~xlsx_table_metrics();
+    xlsx_table_metrics ();
+    ~xlsx_table_metrics ();
 
 public:
-    xlsx_table_position calc(double x_pt, double y_pt);
-    xlsx_table_position calc(size_t offset_col,size_t offset_row,double x_pt, double y_pt);
-	xlsx_table_position calc(size_t last_col,size_t last_row);
+    xlsx_table_position calc (double x_pt, double y_pt);
+    xlsx_table_position calc (int offset_col, int offset_row, double x_pt, double y_pt);
+	xlsx_table_position calc (int last_col, int last_row);
 
-	void update_pt(size_t offset_col,size_t offset_row,double &x_pt, double &y_pt);
+	void update_pt (int offset_col, int offset_row, double &x_pt, double &y_pt);
 
-    void add_cols(size_t count, double widht_pt);
-    void add_rows(size_t count, double height_pt);
+    void add_cols (int count, double widht_pt);
+    void add_rows (int count, double height_pt);
 
 private:
     class Impl;
